+++ packages/openai/src/openai-chat-language-model.test.ts
@@ -176,6 +176,9 @@ describe('doGenerate', () => {
       completion_tokens_details?: {
         reasoning_tokens?: number;
       };
+      prompt_tokens_details?: {
+        cached_tokens?: number;
+      };
     };
     logprobs?: {
       content:
@@ -839,6 +842,33 @@ describe('doGenerate', () => {
     ]);
   });
 
+  it('should return cached_tokens in prompt_details_tokens', async () => {
+    prepareJsonResponse({
+      usage: {
+        prompt_tokens: 15,
+        completion_tokens: 20,
+        total_tokens: 35,
+        prompt_tokens_details: {
+          cached_tokens: 1152,
+        },
+      },
+    });
+
+    const model = provider.chat('gpt-4o-mini');
+
+    const result = await model.doGenerate({
+      inputFormat: 'prompt',
+      mode: { type: 'regular' },
+      prompt: TEST_PROMPT,
+    });
+
+    expect(result.providerMetadata).toStrictEqual({
+      openai: {
+        cachedPromptTokens: 1152,
+      },
+    });
+  });
+
   describe('reasoning models', () => {
     it('should clear out temperature, top_p, frequency_penalty, presence_penalty', async () => {
       prepareJsonResponse();
@@ -909,6 +939,27 @@ describe('doGenerate', () => {
       model: 'o1-preview',
       messages: [{ role: 'user', content: 'Hello' }],
       max_completion_tokens: 255,
+    });
+  });
+
+  it('should send store extension setting', async () => {
+    prepareJsonResponse({ content: '' });
+
+    await model.doGenerate({
+      inputFormat: 'prompt',
+      mode: { type: 'regular' },
+      prompt: TEST_PROMPT,
+      providerMetadata: {
+        openai: {
+          store: true,
+        },
+      },
+    });
+
+    expect(await server.getRequestBodyJson()).toStrictEqual({
+      model: 'gpt-3.5-turbo',
+      messages: [{ role: 'user', content: 'Hello' }],
+      store: true,
     });
   });
 
@@ -1072,6 +1123,9 @@ describe('doStream', () => {
       prompt_tokens: number;
       total_tokens: number;
       completion_tokens: number;
+      prompt_tokens_details?: {
+        cached_tokens?: number;
+      };
     };
     logprobs?: {
       content:
@@ -1678,6 +1732,72 @@ describe('doStream', () => {
       'custom-request-header': 'request-header-value',
       'openai-organization': 'test-organization',
       'openai-project': 'test-project',
+    });
+  });
+
+  it('Should handle cached tokens in experimental_providerMetadata', async () => {
+    prepareStreamResponse({
+      content: [],
+      usage: {
+        prompt_tokens: 15,
+        completion_tokens: 20,
+        total_tokens: 35,
+        prompt_tokens_details: {
+          cached_tokens: 1152,
+        },
+      },
+    });
+
+    const { stream } = await model.doStream({
+      inputFormat: 'prompt',
+      mode: { type: 'regular' },
+      prompt: TEST_PROMPT,
+    });
+
+    expect(await server.getRequestBodyJson()).toStrictEqual({
+      stream: true,
+      stream_options: { include_usage: true },
+      model: 'gpt-3.5-turbo',
+      messages: [{ role: 'user', content: 'Hello' }],
+    });
+
+    const chunksArr = await convertReadableStreamToArray(stream);
+    expect(chunksArr[chunksArr.length - 1]).toHaveProperty('providerMetadata');
+    expect(chunksArr[chunksArr.length - 1].type).toEqual('finish');
+    expect(chunksArr[chunksArr.length - 1]).toStrictEqual({
+      type: 'finish',
+      finishReason: 'stop',
+      logprobs: undefined,
+      usage: {
+        promptTokens: 15,
+        completionTokens: 20,
+      },
+      providerMetadata: {
+        openai: { cachedPromptTokens: 1152 },
+      },
+    });
+  });
+
+  it('should send store extension setting', async () => {
+    prepareStreamResponse({ content: [] });
+
+    await model.doStream({
+      inputFormat: 'prompt',
+      mode: { type: 'regular' },
+      prompt: TEST_PROMPT,
+      providerMetadata: {
+        openai: {
+          store: true,
+        },
+      },
+    });
+
+    expect(await server.getRequestBodyJson()).toStrictEqual({
+      model: 'gpt-3.5-turbo',
+      stream: true,
+      stream_options: { include_usage: true },
+      messages: [{ role: 'user', content: 'Hello' }],
+      store: true,
     });
   });
 });