+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/amazon-bedrock
 
+## 0.0.29
+
+### Patch Changes
+
+- 8c3847e: fix (provider/bedrock): update amazon bedrock package to use safe version of aws sdk
+
 ## 0.0.28
 
 ### Patch Changes